+++ lib/Doctrine/DBAL/Platforms/AbstractPlatform.php
@@ -17,7 +17,7 @@ use Doctrine\DBAL\Platforms\Keywords\KeywordList;
 use Doctrine\DBAL\Schema\Column;
 use Doctrine\DBAL\Schema\ColumnDiff;
 use Doctrine\DBAL\Schema\Constraint;
-use Doctrine\DBAL\Schema\Exceptions\Expression;
+use Doctrine\DBAL\Schema\Expressions\Expression;
 use Doctrine\DBAL\Schema\ForeignKeyConstraint;
 use Doctrine\DBAL\Schema\Identifier;
 use Doctrine\DBAL\Schema\Index;
