+++ tests/Doctrine/Tests/DBAL/Platforms/AbstractPlatformTestCase.php
@@ -10,7 +10,7 @@ use Doctrine\DBAL\Platforms\Keywords\KeywordList;
 use Doctrine\DBAL\Schema\Column;
 use Doctrine\DBAL\Schema\ColumnDiff;
 use Doctrine\DBAL\Schema\Comparator;
-use Doctrine\DBAL\Schema\Exceptions\Expression;
+use Doctrine\DBAL\Schema\Expressions\Expression;
 use Doctrine\DBAL\Schema\ForeignKeyConstraint;
 use Doctrine\DBAL\Schema\Index;
 use Doctrine\DBAL\Schema\Table;
